Update: ResponseDto 수정

+++ src/main/java/com/server/mappin/dto/PostCreateRequestDto.java
@@ -13,7 +13,7 @@ public class PostCreateRequestDto {
     private LocalDate date;
     private String dong;
     private String category;
-    private int x;
-    private int y;
+    private Double x;
+    private Double y;
 
 }
